requirements.txt
fastapi
uvicorn
openai
sqlalchemy
pymysql
jinja2
langchain
boto3
python-dotenv
